Move the required functions from the sparse_merkle_tree dir to another location.
