+++ src/bin/smt_verifier.rs
@@ -1,5 +1,6 @@
-use intmax_zkp_core::sparse_merkle_tree::goldilocks_poseidon::{
-    GoldilocksHashOut, PoseidonSparseMerkleTreeMemory,
+use intmax_zkp_core::{
+    sparse_merkle_tree::goldilocks_poseidon::PoseidonSparseMerkleTreeMemory,
+    utils::hash::GoldilocksHashOut,
 };
 use plonky2::{field::types::Sample, hash::hash_types::HashOut};
