+++ src/lib.rs
@@ -2,11 +2,12 @@ pub mod config;
 #[cfg(feature = "ecdsa")]
 pub mod ecdsa;
 pub mod merkle_tree;
-pub mod poseidon;
+// pub mod poseidon;
 pub mod recursion;
 // pub mod rollup;
 pub mod sparse_merkle_tree;
 pub mod transaction;
+pub mod utils;
 pub mod zkdsa;
 
 pub extern crate plonky2;
