+++ src/sparse_merkle_tree/layered_layered_tree.rs
@@ -5,11 +5,10 @@ use super::{
     node_hash::NodeHash,
     proof::{SparseMerkleInclusionProof, SparseMerkleProcessProof},
     root_data::RootData,
-    tree::{
-        calc_inclusion_proof, calc_process_proof, get, HashLike, KeyLike, SparseMerkleTree,
-        ValueLike,
-    },
+    tree::{calc_inclusion_proof, calc_process_proof, get, SparseMerkleTree},
 };
+
+use crate::merkle_tree::tree::{HashLike, KeyLike, ValueLike};
 
 pub type LayeredLayeredSparseMerkleInclusionProof<K, V, I> = (
     SparseMerkleInclusionProof<K, V, I>,
