+++ src/sparse_merkle_tree/proof/inclusion.rs
@@ -62,7 +62,7 @@ impl<K, V, I> SparseMerkleInclusionProof<K, V, I> {
 
 #[test]
 fn test_serialize_merkle_proof() {
-    use super::super::goldilocks_poseidon::GoldilocksHashOut;
+    use crate::utils::hash::GoldilocksHashOut;
 
     let merkle_proof = SparseMerkleInclusionProof {
         root: GoldilocksHashOut::from_u32(1),
