+++ src/transaction/tree/tx_diff.rs
@@ -8,9 +8,8 @@ use plonky2::{
 use crate::{
     merkle_tree::{
         sparse_merkle_tree::{MerklePath, Node},
-        tree::MerkleProof,
+        tree::{le_bytes_to_bits, MerkleProof},
     },
-    sparse_merkle_tree::goldilocks_poseidon::le_bytes_to_bits,
     transaction::asset::{encode_contributed_asset, ContributedAsset, TokenKind},
     zkdsa::account::Address,
 };
@@ -260,11 +259,11 @@ fn test_prove_tx_diff_tree() {
 
     use crate::{
         merkle_tree::tree::get_merkle_root,
-        sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
         transaction::{
             asset::{TokenKind, VariableIndex},
             tree::tx_diff::TxDiffTree,
         },
+        utils::hash::GoldilocksHashOut,
         zkdsa::account::{private_key_to_account, Address},
     };
 
@@ -329,11 +328,11 @@ fn test_tx_diff_tree_by_plonky2() {
 
     use crate::{
         merkle_tree::tree::{get_merkle_proof, get_merkle_root},
-        sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
         transaction::{
             asset::{TokenKind, VariableIndex},
             tree::tx_diff::TxDiffTree,
         },
+        utils::hash::GoldilocksHashOut,
         zkdsa::account::{private_key_to_account, Address},
     };
