+++ src/transaction/tree/user_asset.rs
@@ -8,9 +8,8 @@ use plonky2::{
 use crate::{
     merkle_tree::{
         sparse_merkle_tree::{MerklePath, Node},
-        tree::MerkleProof,
+        tree::{le_bytes_to_bits, KeyLike, MerkleProof},
     },
-    sparse_merkle_tree::{goldilocks_poseidon::le_bytes_to_bits, tree::KeyLike},
     transaction::asset::{encode_contributed_asset, ContributedAsset, TokenKind, VariableIndex},
     zkdsa::account::Address,
 };
@@ -320,8 +319,8 @@ fn test_prove_user_asset_tree() {
 
     use crate::{
         merkle_tree::tree::get_merkle_root,
-        sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
         transaction::asset::{ContributedAsset, TokenKind, VariableIndex},
+        utils::hash::GoldilocksHashOut,
         zkdsa::account::{private_key_to_account, Address},
     };
 
@@ -395,11 +394,11 @@ fn test_user_asset_tree_by_plonky2() {
 
     use crate::{
         merkle_tree::tree::get_merkle_root,
-        sparse_merkle_tree::goldilocks_poseidon::GoldilocksHashOut,
         transaction::{
             asset::{ContributedAsset, TokenKind, VariableIndex},
             tree::user_asset::UserAssetTree,
         },
+        utils::hash::GoldilocksHashOut,
         zkdsa::account::{private_key_to_account, Address},
     };
